refactor: remove unneeded folder
