+++ src/App.js
@@ -33,7 +33,7 @@ import { InjectAppServices } from './services/pure-di';
 import EditorsDemo from './components/EditorsDemo/EditorsDemo';
 import UpdatePaymentInformation from './components/Plans/UpdatePaymentInformation/UpdatePaymentInformation/index';
 import UpdatePaymentInformationSummary from './components/Plans/UpdatePaymentInformation/UpdatePaymentInformationSummary/index';
-import IntegrationsSection from './components/Integrations/Section/index';
+import IntegrationsSection from './components/Integrations/index';
 
 /**
  * @param { Object } props - props
