+++ src/components/Integrations/images/connected.png
